Fix a bug

+++ src/backend/main.py
@@ -16,8 +16,10 @@ main = Blueprint('main', __name__)
 
 def genQR():
     pattern = "[^0-9a-zA-Z]+"
-    shutil.rmtree("/tmp/medhelper")
-    os.mkdir("/tmp/medhelper")
+    if os.path.isdir(path):
+        shutil.rmtree("/tmp/medhelper")
+    else:
+        os.mkdir("/tmp/medhelper")
     for placement in Placement.query.all():
         img = qrcode.make(f'{url_for("main.app", _external=True)}?uuid={placement.id}')
         img.save(f"/tmp/medhelper/QR_{re.sub(pattern, '_', placement.placement)}.png")
